requirements.txt
numpy~=2.4.2
nibabel~=5.3.3
pandas~=3.0.0
tqdm~=4.67.3
matplotlib~=3.10.8
# xnat~=0.7.2 # only needed for bulk downloading ABIDE through the XNAT channel
python-dotenv~=1.2.1